v1 works
correct lowes, in cart, for pickup, $1.18. Great outcome! Howver, it went to lowes.com, then instead of searching in the lowe's search box, it entered "cheap flat head screwdriver" in the address bar, which performed a google search. It correctly chose a product result from Lowe's and proceeded from there.